+++ include/spatialops/structured/FVStaggeredScratch.h
@@ -134,11 +134,11 @@ namespace structured{
                           std::vector<double> & vals,
                           std::vector<int> & ixs ) const;
 
-    void get_ghost_cols( std::set<int>& ghostCols ) const
+    void get_ghost_cols( std::set<size_t>& ghostCols ) const
     {
       ghostCols = get_ghost_set<SrcFieldT>( dim_, hasPlusXSideFaces_, hasPlusYSideFaces_, hasPlusZSideFaces_ );
     }
-    void get_ghost_rows( std::set<int>& ghostRows ) const
+    void get_ghost_rows( std::set<size_t>& ghostRows ) const
     {
       ghostRows = get_ghost_set<DestFieldT>( dim_, hasPlusXSideFaces_, hasPlusYSideFaces_, hasPlusZSideFaces_ );
     }
